+++ tests/test_converter.py
@@ -1,4 +1,5 @@
 """Unoconverter unit tests"""
+
 import os
 import pytest
 import sys